requirements.txt
torchvision==0.15.1   # needs to be a different version on mac if you want to use the M1 as GPU
matplotlib==3.4.2
lightning==2.0.2
lightning_bolts==0.3.4
numpy==1.20.3
torch==2.0.0
optuna==2.10.0
Pillow==8.4.0
PyYAML==6.0
scikit_learn==1.0.1
torchmetrics==0.11.4
opencv-python
test-tube